+++ src/main/java/duke/DukeException.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * This class is for exceptions specific to the chatbot.
  *
@@ -5,7 +7,7 @@
  */
 public class DukeException extends Exception{
     /**
-     * Constructor for DukeException class.
+     * Constructor for duke.DukeException class.
      *
      * @param message the error message.
      */
