+++ src/main/java/duke/Storage.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.*;
+
 import java.io.*;
 import java.util.ArrayList;
