+++ src/main/java/duke/command/CreateDeadlineCommand.java
@@ -1,13 +1,21 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.io.IOException;
 
-public class CreateTodoCommand extends Command {
+public class CreateDeadlineCommand extends Command {
     private Ui ui;
 
     private TaskList taskList;
 
     private Storage storage;
 
-    public CreateTodoCommand(String commandMessage, Ui ui, TaskList taskList, Storage storage) {
+    public CreateDeadlineCommand(String commandMessage, Ui ui, TaskList taskList, Storage storage) {
         super(commandMessage);
         this.ui = ui;
         this.taskList = taskList;
@@ -17,8 +25,9 @@ public class CreateTodoCommand extends Command {
     @Override
     public boolean execute() {
         try {
-            String[] commandMessageArr = this.commandMessage.split(" ", 2);
-            Task task = new Todo(commandMessageArr[1], false);
+            String[] commandMessageArr = this.commandMessage.split("/", 2);
+            Task task = new Deadline(commandMessageArr[0].substring(9), false,
+                    commandMessageArr[1].substring(3));
             this.taskList.addTask(task);
             this.storage.storeTask(task);
             this.ui.replyTaskAdded(task);
