+++ src/main/java/duke/command/MarkTaskCommand.java
@@ -1,13 +1,20 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.io.IOException;
 
-public class DeleteTaskCommand extends Command {
+public class MarkTaskCommand extends Command {
     private Ui ui;
 
     private TaskList taskList;
 
     private Storage storage;
 
-    public DeleteTaskCommand(String commandMessage, Ui ui, TaskList taskList, Storage storage) {
+    public MarkTaskCommand(String commandMessage, Ui ui, TaskList taskList, Storage storage) {
         super(commandMessage);
         this.ui = ui;
         this.taskList = taskList;
@@ -19,10 +26,10 @@ public class DeleteTaskCommand extends Command {
         try {
             String[] commandMessageArr = this.commandMessage.split(" ", 2);
             int taskNumber = Integer.parseInt(commandMessageArr[1]);
-            Task task = this.taskList.deleteTask(taskNumber);
+            Task task = this.taskList.markTask(taskNumber);
             this.storage.restructure(this.taskList);
-            this.ui.replyTaskDeleted(task);
-        } catch (IOException | DukeException exception) {
+            this.ui.replyTaskMarked(task);
+        } catch (IOException exception) {
             ui.replyError(exception.getMessage());
         }
