+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
@@ -8,12 +10,12 @@ import java.time.format.DateTimeFormatter;
  */
 public class Deadline extends Task {
     /**
-     * Deadline for the task.
+     * duke.task.Deadline for the task.
      */
     private LocalDate by;
 
     /**
-     * Constructor for Deadline class.
+     * Constructor for duke.task.Deadline class.
      *
      * @param desc description of the deadline task.
      */
